Make MaterializedViewTests more stable
